honeypots/requirements.txt
requests
openai
Flask==2.1.3
werkzeug==2.0.3
googlesearch-python



